assembly of works and worksManager files , adding functionality to update the gallery while remembering the selected filter
